Dynamically build schemas and remove caching
