+++ explorer/tasks.py
@@ -78,6 +78,4 @@ def truncate_querylogs(days):
 @task
 def build_schema_cache_async(connection_alias):
     from .schema import build_schema_info, connection_schema_cache_key
-    ret = build_schema_info(connection_alias)
-    cache.set(connection_schema_cache_key(connection_alias), ret)
-    return ret
+    return build_schema_info(connection_alias)
